+++ types/index.d.ts
@@ -493,6 +493,11 @@ declare namespace K {
     mean(x: ReadonlyArray<number>): number;
 
     /**
+     * Takes an array of numbers and calculates the median
+     */
+    median(list: ReadonlyArray<number>): number;
+
+    /**
      * Creates a new function that, when invoked, caches the result of calling fn for a given argument set and returns the result. Subsequent calls to the memoized fn with the same argument set will not result in an additional call to fn; instead, the cached result for that set of arguments will be returned.
      */
     memoizeWith(mFn: Function, fn: Function): Function;