+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   - `addIndex` description tweaks
   - `memoizeWith` added [RamdaJs](https://ramdajs.com) credit since it's based mostly off their version
   - `lt`, `lte`, `gt`, and `gte` all had invalid examples in their documentation. This is addressed now
+  - `find` and `deepEq` examples were incorrect
 
 ## v0.11.3