+++ CHANGELOG.md
@@ -15,7 +15,9 @@
 - Added a `mod` function which behaves like a modulo would in actual mathematics
   - Added onto the types definitions
 - Added `somePass` function which takes an array of functions and passes a value through them until one passes or the end is reached
-  - Added into the types def
+  - Added onto the types definitions
+- Added `everyPass` function which takes an array of functions and passes a value through until one fails or the end is reached
+  - Added onto the types definitions
 
 ### Improved